🎉 NEW FEATURE: Add game status field to in memory stored session
